+++ custom_model_runner/datarobot_drum/drum/language_predictors/r_common_code/common.R
@@ -4,7 +4,8 @@ TargetTypeEnum <- function() {
       REGRESSION = "regression",
       BINARY = "binary",
       ANOMALY = "anomaly",
-      UNSTRUCTURED = "unstructured"
+      UNSTRUCTURED = "unstructured",
+      MULTICLASS = "multiclass"
     )
 }