+++ custom_model_runner/CHANGELOG.md
@@ -5,14 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+#### Unreleased
+##### Added
+- add `multiclass` target type
+- add `class-labels` for specifying the class labels of a multiclass model
+- add `class-labels-file` for specifying a file containing the class labels of a multiclass model
+
 #### [1.3.0] - 2020-10-15
 ##### Added
 - `read_input_data` hook
 - add `target-type` param
 - unstructured mode
-- add `multiclass` target type
-- add `class-labels` for specifying the class labels of a multiclass model
-- add `class-labels-file` for specifying a file containing the class labels of a multiclass model
 
 #### [1.2.0] - 2020-08-28
 ##### Added